+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 Änderungen
 ----------
 v1.0.14 (??.??.2016)
+ * unused getSelector method in be lister removed
  * new decorator interface and base class added
  * initialized parameter object in be modules added (`$mod->getConfigurations()->getParameters()->getInt('uid')` returns `5` from url `?SET['uid']`)
 
 v1.0.13 (09.06.2016)
- + php support lower then 5.4 restored in be module dispatcher
+ * php support lower then 5.4 restored in be module dispatcher
  * fallbackPid check for be modules added
  * action url for be modules fixed